backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
playwright
psycopg2-binary